requirements.txt
flask
gunicorn
jsonschema
requests
pyyaml